completes support of jquery and vue in _card_use.html
Clicking update on /billing/{organization}/card when
the `document.referrer` is `/login/` would return to
the login page.
Removing the `document.referrer` in the redirect does
not permit to update the card-on-file in the /billing/cart/
page while at the same time carrying a redirect after
the payment is completed.

The best way to resolve the issue is to stay on the
/billing/cart/ page when a credite-card-on-file is updated.