retrieves provider.processor_backend only
This is a follow-on to commit c38b6b43.

When using StripeConnect, a card will be stored
on a provider's or the broker's account.
A subscriber technically has no Stripe account
and if we were to use its `processor_backend`,
incorrect keys would be set when calling the Stripe API.