fixes regression in commit 45fcc55f
commit 9d3fd405 actually introduced the `or` instead of `and` operator
so we can retrieves detailed information on specific profiles in a single
API call. Example:

`GET /api/profile/cowork/plans/medium/subscriptions?q_f=profile&q=xia23,xia24`

This commit thus restricts separators to commas instead of commas and spaces.
As a result quotes (") are unnecessary - which is what we were looking for.