creates implicit personal profile if hint is present in plan
Multiple sites enable to subscribe either a personal profile or
organization profile without a clear progression from one to the other.
i.e. some users will start with a role on an organization, then subscribe
to a personal plan, while other will start with a personal profile
and connect to an organization role afterwards.

A basic `implicit_create_on_none` is not enough to support onboarding
of both types. We add a hint flag (`is_personal`) on `Plan` models
to give more flexibility to the onboarding flow.