verifies request.user e-mail whenever we attempt an implicit grant
Before, we would only verify the e-mail address if an implicit
grant was about to be performed. Now even if no grant is performed,
we still check the e-mail address on /users/roles/accept/.

Note that we are still not doing any e-mail verification if there
is a role associated to the user to maintain frictionless registration
when a personal profile is created.