re-adds `created_at` in /api/metrics/registered
This API end point, only available to brokers, is used to identify
users that have registered but which are not connected to a subscription.
Having the date_joined is thus important here.

Also in this commit, calls the registered API lazily in the Vue
component because when there is over 300K users in the database
the computation of the queryset is a bit slow on m3.medium.