accounts for email in different case in group buy
`bulk_items` in `execute_order` would return an empty set if the profile
email in the database had a different case than the one in the `Cart.sync_on`
field, leading to the else condition and a MultipleObject exception.

Note `translate_descr_suffix` was also updated such that the group buy
benefiary doesn't get dropped in translation.