keeps middle name in user.last_name when inviting a user
Before that commit, we relied on `full_name_natural_split` to
return first, middle and last names, then dropped the middle name
when creating the user associated with the invite. As a result
the reconstructed `full_name` presented on activation of the account
would not match the `full_name` entered in the invite.